refactor: 패키지 의존성 순환참조 문제를 해결
