+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -2,6 +2,8 @@ package blackjack.domain.participant;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Deck;
+import blackjack.domain.card.Score;
+import blackjack.domain.state.Hand;
 import blackjack.domain.state.State;
 import java.util.List;
