+++ src/main/java/blackjack/domain/state/BustState.java
@@ -1,7 +1,5 @@
 package blackjack.domain.state;
 
-import blackjack.domain.participant.Hand;
-
 public class BustState extends ClosedState {
 
     BustState(Hand hand) {
