+++ src/main/java/blackjack/domain/state/HitState.java
@@ -2,7 +2,6 @@ package blackjack.domain.state;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Deck;
-import blackjack.domain.participant.Hand;
 
 public class HitState extends State {
