+++ src/test/java/blackjack/domain/participant/PlayersTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 import blackjack.domain.card.BlackjackCardFactory;
 import blackjack.domain.card.Deck;
+import blackjack.domain.card.Score;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
